+++ CHANGELOG.md
@@ -33,7 +33,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
     > `?filtering={$OR;lastname:eq:foo;firstname:eq:foo};{$OR;age:gt:30;age:lt:20}` → (lastname == "foo" OR firstname == "foo") AND (age > 30 OR age < 20)
   - within the value(s) of a condition semicolons and braces (i.e. `;{}`) must be backslash escaped; URL encoding is necessary as usual
 
-- Added Lobid-Clients for missing Lobid objects
+- Add Lobid-Clients for missing Lobid objects
+- Add persistence support for new field `Headword.labelNormalized`
+- Add sorting to Buckets and Bucket objects handling in `HeadwordRepositoryImpl` 
 
 ### Changed
 
@@ -52,6 +54,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
   - Switch language of displayed multilanguage data fields now using a select drop down instead of tabs in view and list pages
   - Switch language of displayed multilanguage data fields now also handling languages with script
 - Thymeleaf Date and TimeValue rendering fragment (date.html) and messages properties moved from cudami admin webapp to cudami client for easy reuse in frontend clients
+- Separate `PageRequest` param handling from `ListRequest` param handling in `BaseRestClient`
+- Reuse logic from `BaseRestClient` in `CudamiHeadwordsClient`
 
 ### Fixed
 
@@ -59,6 +63,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/) 
 - **SQL migration** (can be **long running**) replaces every emtpy string key in the labels (that comes from `Locale.ROOT` as language)
 by `"und"`
 - `getByValue` method in `PredicateRepositoryImpl`
+- Comment not supported sorting throwing warning every time in `EntityToEntityRelationRepositoryImpl`
 
 ### Removed